httpmessage_message_content_consume() honor Content-Length header
* Add httpmessage_int_consume()
  Parse an integer text representation
* httpmessage_message_consume() evaluates Content-Length header field value if any and adjust body length accordingly.
* Add unit test to verify the new features

+++ src/httpmessage/header.c
@@ -442,7 +442,7 @@ int httpmessage_headerfield_line_consume(
 			return result;
 		}
 		
-		return (consumed + (size_t)result);
+		return (int)(consumed + (size_t)result);
 	}
 	
 	/* Header field name */
